+++ src/utils/variable.test.ts
@@ -2,7 +2,7 @@ import { locationService } from '@grafana/runtime';
 
 import { ALL_VALUE, ALL_VALUE_PARAMETER, NO_VALUE_PARAMETER } from '../constants';
 import { VariableChangedEvent, VariableType } from '../types';
-import { selectVariableValues } from './variable';
+import { getRuntimeVariable, selectVariableValues } from './variable';
 
 /**
  * Mock @grafana/runtime
@@ -159,6 +159,34 @@ describe('Variable Utils', () => {
         expect(locationService.partial).toHaveBeenCalledWith(
           {
             [`var-${variable.name}`]: ['value1', 'value2', 'selected1', 'selected2'],
+          },
+          true
+        );
+      });
+
+      it('Should remove "$__all" values from already selected', () => {
+        jest.mocked(locationService.getSearch).mockImplementation(
+          () =>
+            ({
+              getAll: jest.fn(() => [ALL_VALUE_PARAMETER]),
+            }) as any
+        );
+        const variable = {
+          name: 'variable',
+          type: VariableType.CUSTOM,
+          current: { value: [ALL_VALUE_PARAMETER] },
+          options: [],
+          multi: true,
+        };
+        selectVariableValues({
+          values: ['value1', 'value2'],
+          runtimeVariable: variable as any,
+          panelEventBus: eventBus,
+        });
+
+        expect(locationService.partial).toHaveBeenCalledWith(
+          {
+            [`var-${variable.name}`]: ['value1', 'value2'],
           },
           true
         );
@@ -473,24 +501,6 @@ describe('Variable Utils', () => {
       });
     });
 
-    describe('Constant', () => {
-      it('Should apply only first value', () => {
-        const variable = { name: 'variable', type: VariableType.CONSTANT };
-        selectVariableValues({
-          values: ['value1', 'value2'],
-          runtimeVariable: variable as any,
-          panelEventBus: eventBus,
-        });
-
-        expect(locationService.partial).toHaveBeenCalledWith(
-          {
-            [`var-${variable.name}`]: 'value1',
-          },
-          true
-        );
-      });
-    });
-
     it('Should not apply value if no variable passed', () => {
       selectVariableValues({ values: ['value1'], panelEventBus: eventBus });
 
@@ -502,6 +512,64 @@ describe('Variable Utils', () => {
       selectVariableValues({ values: ['value1'], runtimeVariable: variable as any, panelEventBus: eventBus });
 
       expect(locationService.partial).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('getRuntimeVariable', () => {
+    it('Should return variable with "All" option if option not return in options list', () => {
+      const variable = {
+        name: 'variable',
+        type: VariableType.CUSTOM,
+        current: {
+          value: ALL_VALUE_PARAMETER,
+        },
+        options: [
+          {
+            text: 'value1',
+            value: 'value1',
+            selected: false,
+          },
+          {
+            text: 'value2',
+            value: 'value2',
+            selected: false,
+          },
+        ],
+        includeAll: true,
+        multi: true,
+      } as any;
+      const result = getRuntimeVariable(variable);
+
+      expect(result?.options[0].value).toEqual(ALL_VALUE_PARAMETER);
+      expect(result?.options[0].selected).toBeTruthy();
+    });
+
+    it('Should return variable with "All" option if option not return in options list and not select if not present in current value', () => {
+      const variable = {
+        name: 'variable',
+        type: VariableType.CUSTOM,
+        current: {
+          value: [],
+        },
+        options: [
+          {
+            text: 'value1',
+            value: 'value1',
+            selected: false,
+          },
+          {
+            text: 'value2',
+            value: 'value2',
+            selected: false,
+          },
+        ],
+        includeAll: true,
+        multi: true,
+      } as any;
+      const result = getRuntimeVariable(variable);
+
+      expect(result?.options[0].value).toEqual(ALL_VALUE_PARAMETER);
+      expect(result?.options[0].selected).not.toBeTruthy();
     });
   });
 });