+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated behavior in table view for Textbox variable type (#211)
 - Added Date Time Selector (#209)
+- Add All options in Scenes dashboards (#213)
 
 ## 3.5.0 (2024-10-02)